# policy: Do not scan all entries when adding new ones