--- a/pkg/policy/mapstate.go
+++ b/pkg/policy/mapstate.go
@@ -98,6 +98,7 @@ type MapState interface {
 	// For testing from other packages only
 	Equals(MapState) bool
 	Diff(expected MapState) string
+	WithState(initMap map[Key]MapStateEntry) MapState
 }
 
 // mapState is a state of a policy map.
@@ -340,11 +341,22 @@ func getNets(identities Identities, ident uint32) []netip.Prefix {
 }
 
 // NewMapState creates a new MapState interface
-func NewMapState(initMap map[Key]MapStateEntry) MapState {
-	return newMapState(initMap)
+func NewMapState() MapState {
+	return newMapState()
 }
 
-func newMapState(initMap map[Key]MapStateEntry) *mapState {
+func (ms *mapState) WithState(initMap map[Key]MapStateEntry) MapState {
+	return ms.withState(initMap)
+}
+
+func (ms *mapState) withState(initMap map[Key]MapStateEntry) *mapState {
+	for k, v := range initMap {
+		ms.insert(k, v)
+	}
+	return ms
+}
+
+func newMapState() *mapState {
 	m := &mapState{
 		allows: mapStateMap{
 			entries: make(map[Key]MapStateEntry),
@@ -354,9 +366,6 @@ func newMapState(initMap map[Key]MapStateEntry) *mapState {
 			entries: make(map[Key]MapStateEntry),
 			trie:    bitlpm.NewTrie[Key, map[identity.NumericIdentity]struct{}](policyTypes.MapStatePrefixLen),
 		},
-	}
-	for k, v := range initMap {
-		m.insert(k, v)
 	}
 	return m
 }
@@ -1086,7 +1095,7 @@ func (ms *mapState) authPreferredInsert(newKey Key, newEntry MapStateEntry, feat
 			// New entry has a default auth type.
 			// Fill in the AuthType from more generic entries with an explicit auth type
 			maxSpecificity := 0
-			l3l4State := newMapState(nil)
+			l3l4State := newMapState()
 
 			ms.ForEachAllow(func(k Key, v MapStateEntry) bool {
 				// Only consider the same Traffic direction