Hold MapStateEntries in `entries` map, `trie` and `cidr` within the `trie` are additional indices for the same MapStateEntries. `cidr` only indexes keys with CIDR identities, while `trie` indexes keys by traffic direction, protocol, destination port and mask.

Use the `trie` and the new `cidr` index when adding entries to MapState with deny entries, and the existing `trie` index when adding entries to MapState with auth entries. Likewise, use the `trie` index when adding visibility entries. With this we no longer scan through all entries in MapState when adding new entries.

MapState now contains an optional validator interface that can be set in unit tests to add calls to validation code for the new ancestor/descendants code. This is currently enabled only for the `denyPreferredInsertWithChanges`.

On local benchmarking for generating MapState for an egress policy with 1000 deny and 1000 allow CIDR rules:

Before:
```
Number of MapState entries: 117515
goos: linux
goarch: arm64
pkg: github.com/cilium/cilium/pkg/policy
BenchmarkRegenerateCIDRDenyPolicyRules-12    	       1	763361909494 ns/op
PASS
ok  	github.com/cilium/cilium/pkg/policy	763.696s
```
After:
```
Number of MapState entries: 117515
goos: linux
goarch: arm64
pkg: github.com/cilium/cilium/pkg/policy
BenchmarkRegenerateCIDRDenyPolicyRules-12    	Number of MapState entries: 117515
Number of MapState entries: 117515
       3	 418981451 ns/op
PASS
ok  	github.com/cilium/cilium/pkg/policy	3.434s
```
This is a 1820x speedup :-)

Using the PortProto `trie` in addition to the new `cidr` trie as suggested by @nathanjsweet made this 8x faster for this specific test case.

```release-note
Deny rule processing has been improved to run faster, especially on larger policies.
```