@jrajahalme 2024-06-22
Moved identityIsSupersetOf validation from production code to unit test. This sped up the bechmark almost 2x and removed panics from production code.