@jrajahalme 2024-06-22
Split refactoring from commit `policy: Add cidr index to mapstate` to a new commit just before it.