@jrajahalme 2024-06-26
> A better approach would be to index the CIDRs by key and do simple lookups within a port-protocol traversal.

Interesting, could you elaborate on this, i.e., be more specific?