@jrajahalme 2024-06-26
> It is also unclear to me why this PR traverses all CIDR ancestors of a given policy identity, when most of them can be completely eliminated from comparison by a simple port-protocol traversal of the indices.

In any case the CIDR ancestors is a small subset of all entries, as all (other) leaf entries as well as all non-CIDR entries are eliminated.

The problem with proto-protocol traversal is that it could degenerate to scanning almost all entries, if there are many CIDR rules for a few port-protocol pairs or wildcards thereof.