@jrajahalme 2024-06-26
> This PR is on the right track, but how CIDRs and identity are indexed to one another and compared needs overhaul. This PR optimizes CIDR comparisons, but at the expense of the gains that can be made by optimizing port protocol comparison, which is, by far, the more common case. These two things do not have to be mutually exclusive, but the insertion logic in this PR neglects the gains that could be made with more robust port-protocol comparison. It is also unclear to me why this PR traverses all CIDR ancestors of a given policy identity, when most of them can be completely eliminated from comparison by a simple port-protocol traversal of the indices. A better approach would be to index the CIDRs by key and do simple lookups within a port-protocol traversal.

So the design I came up with adds a separate CIDR index, that needs to find a subset of key/entry pairs. I did consider changing MapState to hold MapStateEntry pointers instead of entries themselves to be able to point to the same entries from multiple indices. It seemed simpler to just store the full Keys in the indices and have the entries in a separate map. This makes some operations on the mapstate also faster (getting a specific entry with full key, iterating all entries).

CIDR index provides the "identityIsSuperset" relations almost directly, and I saw that as the potential main determinant of performance, as most identities in the MapState would not satisfy that relation. I left the existing port/protocol checks as-is. I did not try stacking the indices see how much it would help.