@nathanjsweet 2024-06-26
@jrajahalme 
> I did not try stacking the indices see how much it would help.

I think the port-protocol indices should be used first. This makes the logic easier to follow in insertion.

[Here is a draft PR ](https://github.com/cilium/cilium/pull/33413) that modifies this one up to the identity comparison logic using CIDR cache (so the whole PR isn't represented).