@jrajahalme 2024-06-28
> I think the port-protocol indices should be used first. This makes the logic easier to follow in insertion.

@nathanjsweet Thanks for this push-back! I implemented this into this PR, and adding the use of the PortProto `trie` in insertion made the new CIDR-specific benchmark 8x faster, to a combined 1800x speedup!