@jrajahalme 2024-06-28
Rebased to pick up merged  #33447.

`cidr` can be `nil`, had to add check for it to avoid `nil` dereference.