@jrajahalme 2024-07-06
Rebased to fix merge conflicts